+++ Changelog.md
@@ -1,6 +1,12 @@
+## New features
+
+- Can edit group event details (issue #221)
+
+
 # Version 1.5.1 (25 March 2016)
 
 ## New features
+
 - Can create projects and groups from top menu.
 - Send weekly email notifications in groups with updated bibliography recomendations.
 - Better styling of bibliography items in group section.